alteração imagens

+++ GrandeOrvalho_db_django/GrandeOrvalho/models/animais.py
@@ -9,7 +9,7 @@ from .raca import Raca
 class Animais(models.Model):
     SEXO_CHOICES = (("F", "Feminino"), ("M", "Masculino"), ("N", "Nenhuma das opções"))
     nome = models.CharField(max_length=50)
-    idade = models.IntegerField()
+    idade = models.DecimalField(max_digits=10, decimal_places=2)
     peso = models.DecimalField(max_digits=5, decimal_places=2)
     sexo = models.CharField(max_length=1, choices=SEXO_CHOICES, blank=False, null=False)
     status = models.CharField(max_length=20)
